move nats publisher to its own crate
