+++ Cargo.toml
@@ -61,6 +61,7 @@ fuel-core-chain-config = { version = "0.26.0", path = "./crates/chain-config", d
 fuel-core-client = { version = "0.26.0", path = "./crates/client" }
 fuel-core-database = { version = "0.26.0", path = "./crates/database" }
 fuel-core-metrics = { version = "0.26.0", path = "./crates/metrics" }
+fuel-core-nats = { version = "0.26.0", path = "./crates/fuel-core-nats" }
 fuel-core-services = { version = "0.26.0", path = "./crates/services" }
 fuel-core-consensus-module = { version = "0.26.0", path = "./crates/services/consensus_module" }
 fuel-core-bft = { version = "0.26.0", path = "./crates/services/consensus_module/bft" }
