+++ crates/fuel-core-nats/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "fuel-core-nats"
+version = "0.26.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.86"
+async-nats = "0.35.1"
+fuel-core-types = { workspace = true }
+serde_json = "1.0.117"
+tokio = "1.37.0"
+tracing = "0.1.40"
